pyproject.toml
[build-system]
requires = [
    "setuptools>=80.9.0",
    "wheel>=0.45.1"
]
build-backend = "setuptools.build_meta"

[project]
name = "document_processor"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.0",
    "pydantic",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "python-dotenv",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
]

[project.optional-dependencies]
test = [
    "pytest-asyncio",
]
dev = [
    "pyright",
]

[tool.setuptools.packages.find]
where = ["lib"]

[project.scripts]
deploy-backend = "backend_placeholder.server:serve"

[tool.pytest.ini_options]
addopts = "-v"
